Refactored contexts and hooks to their own folders
